reorganized test files
