+++ lib/terminal/test/application/browserUtilities/machines.ts
@@ -1,4 +1,4 @@
-/* lib/terminal/test/application/browser_machines - An object expressing a map of the various remote machines. */
+/* lib/terminal/test/application/browserUtilities/machines - An object expressing a map of the various remote machines. */
 
 import { cpus } from "os";
