+++ lib/terminal/test/application/browserUtilities/mainMenu.ts
@@ -1,4 +1,4 @@
-/* lib/terminal/test/application/browser_mainMenu - A convenience function that opens the main menu while in browser tests. */
+/* lib/terminal/test/application/browserUtilities/mainMenu - A convenience function that opens the main menu while in browser tests. */
 
 const mainMenu = function terminal_test_samples_mainMenu(machine:string):testBrowserItem {
     return {
