+++ lib/terminal/test/application/browserUtilities/showContextMenu.ts
@@ -1,4 +1,4 @@
-/* lib/terminal/test/application/browser_showContextMenu - A convenience function that launches the modal context menu in browser tests. */
+/* lib/terminal/test/application/browserUtilities/showContextMenu - A convenience function that launches the modal context menu in browser tests. */
 
 const showContextMenu = function terminal_test_samples_showContextMenu(node:testBrowserDOM, test:testBrowserTest[], machine:string):testBrowserItem {
     return {
