+++ lib/terminal/test/samples/browser_user.ts
@@ -1,5 +1,5 @@
 
-/* lib/terminal/test/samples/browser_device - A list of tests that execute in the web browser and require multiple computers. */
+/* lib/terminal/test/samples/browser_user - A list of tests that execute in the web browser and require multiple computers. */
 
 const browserUser:testBrowserItem[] = [];
