+++ lib/terminal/test/samples/readme.md
@@ -5,7 +5,7 @@ The files that contain test samples and any sample type specific utilities.
 <!-- Do not edit below this line.  Contents dynamically populated. -->
 
 * **[browser_device.ts](browser_device.ts)** - A list of tests that execute in the web browser and require multiple computers.
-* **[browser_device.ts](browser_device.ts)** - A list of tests that execute in the web browser and require multiple computers.
-* **[browser_self.ts](browser_self.ts)**     - A list of tests that execute in the web browser only on this computer.
-* **[service.ts](service.ts)**               - A list of service tests.
-* **[simulation.ts](simulation.ts)**         - A list of command related tests for running shell simulations against the supported commands.
+* **[browser_self.ts](browser_self.ts)** - A list of tests that execute in the web browser only on this computer.
+* **[browser_user.ts](browser_user.ts)** - A list of tests that execute in the web browser and require multiple computers.
+* **[service.ts](service.ts)**           - A list of service tests.
+* **[simulation.ts](simulation.ts)**     - A list of command related tests for running shell simulations against the supported commands.
